pyproject.toml
[tool.poetry]
name = "hero-api"
version = "0.1.0"
description = ""
authors = ["bilalmk <bilalmk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
mysqlclient = "^2.2.4"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
